+++ packages/patternfly-4/react-core/src/components/OptionsMenu/OptionsMenu.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import styles from '@patternfly/react-styles/css/components/OptionsMenu/options-menu';
-import { css } from '@patternfly/react-styles';
 import { DropdownContext } from '../Dropdown';
 import { DropdownWithContext } from '../Dropdown/DropdownWithContext';
 
@@ -15,13 +14,13 @@ export enum OptionsMenuDirection {
 }
 
 export interface OptionsMenuProps extends React.HTMLProps<HTMLDivElement> {
-  /** Classes applied to root element of the Options menu */
+  /** Classes applied to root element of the options menu */
   className?: string;
-  /** Id of the root element of the Options menu */
+  /** Id of the root element of the options menu */
   id: string;
-  /** Array of OptionsMenuItem and/or OptionMenuItemGroup nodes that will be rendered in the Options menu list */
+  /** Array of OptionsMenuItem and/or OptionMenuGroup nodes that will be rendered in the options menu list */
   menuItems: React.ReactNode[];
-  /** Either an OptionsMenuToggle or an OptionsMenuToggleWithText to use to toggle the Options menu */
+  /** Either an OptionsMenuToggle or an OptionsMenuToggleWithText to use to toggle the options menu */
   toggle: React.ReactElement;
   /** Flag to indicate the toggle has no border or background */
   isPlain?: boolean;
@@ -31,11 +30,11 @@ export interface OptionsMenuProps extends React.HTMLProps<HTMLDivElement> {
   isText?: boolean;
   /** Flag to indicate if menu is groupped */
   isGrouped?: boolean;
-  /** Provides an accessible name for the Options menu */
+  /** Provides an accessible name for the options menu */
   ariaLabelMenu?: string;
   /** Indicates where menu will be aligned horizontally */
   position?: 'right' | 'left';
-  /** Menu will open up or open down from the Options menu toggle */
+  /** Menu will open up or open down from the options menu toggle */
   direction?: 'up' | 'down';
 }
 
@@ -60,7 +59,7 @@ export const OptionsMenu: React.FunctionComponent<OptionsMenuProps> = ({
       toggleClass: isText ? styles.optionsMenuToggleButton : styles.optionsMenuToggle,
       baseClass: styles.optionsMenu,
       disabledClass: styles.modifiers.disabled,
-      menuComponent: 'ul',
+      menuComponent: isGrouped ? 'div' : 'ul',
       baseComponent: 'div'
     }}
   >