+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.41](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.11.40...@patternfly/react-topology@2.11.41) (2019-12-13)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.11.40](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.11.39...@patternfly/react-topology@2.11.40) (2019-12-12)
 
 **Note:** Version bump only for package @patternfly/react-topology